requirements.txt
flask==2.3.3
openai==1.79.0
python-dotenv==1.0.0
langchain==0.3.25
langchain_openai==0.3.17
flask-wtf==1.2.1
email-validator==2.1.0
flask-mail==0.9.1
requests==2.31.0
werkzeug==2.3.7
apscheduler==3.11.0
tzlocal==5.3.1
tiktoken==0.9.0
flask-limiter==3.5.0
flask-talisman==1.1.0
flask-compress==1.14
gunicorn==21.2.0
python-jose==3.3.0
bcrypt==4.0.1
cryptography==41.0.7
pillow==10.1.0
python-magic==0.4.27


